fix(tui): size db result table to the blocks-view region height
